+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 
 The v1 release supports Cisco IOS-XR release versions from 7.7.1 to 7.9.1.
 
+### v1.1.12 (2023-06-30)
+
+- Add Docker Compose v2 support to `host-check` and `xr-compose` scripts.
+
 ### v1.1.11 (2023-06-23)
 
 - Add a check in `host-check` to verify the correct kernel modules parameters are being used.